+++ src/app/dashboard/(admin)/review-applications/[campaignId]/page.tsx
@@ -13,16 +13,16 @@ import useUpdateCampaign from "@/src/hooks/campaigns/useUpdateCampaign";
 import useCreateFinalAnswer from "@/src/hooks/answers/useCreateFinalAnswer";
 import { Button, Fab } from "@mui/material";
 import { Check, Close, Replay } from "@mui/icons-material";
-import CampaignInformationSection from "@/src/components/ongoing-campaigns/edit/CampaignInformationSection";
-import CampaignMediaSection from "@/src/components/ongoing-campaigns/edit/CampaignMediaSection";
-import ContactInformationSection from "@/src/components/ongoing-campaigns/edit/ContactInformationSection";
-import EditCampaignActions from "@/src/components/ongoing-campaigns/edit/EditCampaignActions";
-import EditCampaignDialogs from "@/src/components/ongoing-campaigns/edit/EditCampaignDialogs";
-import EditCampaignHeader from "@/src/components/ongoing-campaigns/edit/EditCampaignHeader";
-import GardenInformationSection from "@/src/components/ongoing-campaigns/edit/GardenInformationSection";
-import GardenStorySection from "@/src/components/ongoing-campaigns/edit/GardenStorySection";
+import CampaignInformationSection from "@/src/components/approved-campaigns/edit/CampaignInformationSection";
+import CampaignMediaSection from "@/src/components/approved-campaigns/edit/CampaignMediaSection";
+import ContactInformationSection from "@/src/components/approved-campaigns/edit/ContactInformationSection";
+import EditCampaignActions from "@/src/components/approved-campaigns/edit/EditCampaignActions";
+import EditCampaignDialogs from "@/src/components/approved-campaigns/edit/EditCampaignDialogs";
+import EditCampaignHeader from "@/src/components/approved-campaigns/edit/EditCampaignHeader";
+import GardenInformationSection from "@/src/components/approved-campaigns/edit/GardenInformationSection";
+import GardenStorySection from "@/src/components/approved-campaigns/edit/GardenStorySection";
 import {DEFAULT_CAMPAIGN_DATA, EditCampaignFormData, TextFieldKey,} from "@/src/types/frontend/campaignEdit";
-import {beneficiaryOptions, categoryOptions,} from "@/src/components/ongoing-campaigns/options";
+import {beneficiaryOptions, categoryOptions,} from "@/src/components/approved-campaigns/options";
 import {useCampaignEditData} from "@/src/hooks/campaigns/useCampaignEditData";
 import BaseModal from "@/src/components/bases/BaseModal";
 import BaseAlert from "@/src/components/bases/BaseAlert";
