+++ src/components/approved-campaigns/edit/CampaignInformationSection.tsx
