+++ src/components/approved-campaigns/edit/GardenInformationSection.tsx
